+++ src/main/kotlin/com/dgu/prompt/blaybus_backend/data/dto/NotificationRequest.kt
@@ -6,7 +6,7 @@ import com.dgu.prompt.blaybus_backend.data.entity.NotificationType
 data class NotificationRequest(
     val employeeNumber: Int,
     val type: NotificationType,
-    val title: String = "",
+    val title: String = "알림",
     val points: Int = 0,
     val postTitle: String = "",
     val period: String = "" // "weekly", "monthly" etc.
